\name{ensolomas-package}
\alias{ensolomas-package}
\alias{ensolomas}
\docType{package}
\title{\packageTitle{ensolomas}}
\description{\packageDescription{ensolomas}}
\section{Package Content}{\packageIndices{ensolomas}}
\author{\packageAuthor{ensolomas}}
\section{Maintainer}{\packageMaintainer{ensolomas}}
\keyword{package}
