% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enso_lomas_lib.R
\name{load_group_homerange_data}
\alias{load_group_homerange_data}
\title{Load Group Homerange Data}
\usage{
load_group_homerange_data(path)
}
\arguments{
\item{path}{to file output by Odd's ctmm pipeline}
}
\value{
data.frame of loaded data
}
\description{
Load Group Homerange Data
}
